CHROMIUM: Disable VMX unless directed not to in the command line
With this patch, the kernel will by default disable VMX on all cores.
This is an improvement over what we had with earlier
systems, which only disabled it only on core 0.

There is a command line parameter called disablevmx.
If it is not specified in the command line VMX will be disabled.
If it is specified with no value, VMX will be disabled.
If it is specified with the value 'on', VMX will be disabled.
If it is specified with a value other than 'off', or 'on', the message
Malformed early option 'disablevmx'
will appear in the log, and VMX will *still* be disabled.

If the value is 'off' VMX will not be disabled.
At the same time, the kernel will
not specifically enable it. It takes no action, leaving
the management decision for later.

This has been tested on a 64-bit platform with and without the option,
with values (both incorrect and correct) for the option, and it works.
I don't have a 32-bit platform to test on.

Without the disablevmx=off option the kernel prints this:
[    0.000000] Disabling VMX on cpu 0
[    0.139521] Disabling VMX on cpu 1
[    0.153338] Disabling VMX on cpu 2
[    0.167148] Disabling VMX on cpu 3

BUG=chrome-os-partner:10449

TEST=build an image for 64-bit Chromeos platform as normal. Boot with as
normal and observe this output:

[    0.000000] Disabling VMX on cpu 0
[    0.139521] Disabling VMX on cpu 1
[    0.153338] Disabling VMX on cpu 2
[    0.167148] Disabling VMX on cpu 3

Add the do_not_disablevmx flag from the command line. There is no output,
but rdmsr commands for register 3a will show 0x0000000000000000, e.g.
localhost ~ # rdmsr 0 0x3a
0x0000000000000000
localhost ~ # rdmsr 1 0x3a
0x0000000000000000
localhost ~ # rdmsr 2 0x3a
0x0000000000000000
localhost ~ # rdmsr 3 0x3a
0x0000000000000000

UNLESS, on, e.g. a lumpy, the lock bit was set, in which case
the msrs will have whatever value they had last boot time.

We had a discussion with a kernel developer about upstreaming this option
and we did not reach closure. I'm assuming we'll be owning this patch
for a while, so I'd like to limit the footprint to one file.

2nd commit message:

CHROMIUM: vmx: control vmx correctly on resume

On x86 chromeos systems with vmx support, the feature is controlled by
the disablevmx flag. There are several boot paths for the CPUs
and this patch ensures that the resume path for the BSP is covered.

Rename cpu_disable_vmx to cpu_control_vmx. It's a bit more
descriptive since it does not disable vmx in all cases.
Further, have the function itself test the variable; this
is more in keeping with how other such capabilities
are controlled anyway and it keeps the variable in
file scope. Finally, make it all conditional on
CONFIG_CHROMEOS.

We use the variable to control the enable/disable instead
of saved CPU context. I could add the FEATURES MSR
to the saved context and use that on restore but
that adds more changes, and I've seen no evidence that anyone
outside chromeos is interested in disabling VMX, so it
makes more sense to me to limit the footprint of this
patch.

3rd commit message:

CHROMIUM: Enable VMX if we are not disabling it.

Previously, we used the disablevmx flag to mean 'disable if not set or
if set to on, do nothing if set to off'. However, since we do not whitelist
the register (0x3c) that would let programs enable vmx, this left us
stranded, with a non-disabled vmx we could not enable. It's simplest
to just enable it and lock it. Only test kernels are built for vmx,
so enabling it is really not going to get anyone very far.

The new meaning:
1. if disablevmx is not set, disable vmx and lock the register
2. if disablevmx is set to on, disable vmx and lock the register
3. if disablevm is set to off, enable vmx and lock the register

Change-Id: I2aa6f258ffcc04f8d146f878b19f13a805d64f0b
Signed-off-by: Ronald G. Minnich <rminnich@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26688
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Will Drewry <wad@chromium.org>

Conflicts:
	arch/x86/kernel/cpu/common.c
[filbranden: 3.18 rebase. Solved trivial conflicts.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>

[rebase44(groeck): Resolved minor conflicts; updated subject;
	squashed three related patches into one]
Signed-off-by: Guenter Roeck <groeck@chromium.org>

Conflicts:
	arch/x86/power/cpu.c
	arch/x86/kernel/cpu/common.c

[rebase412(groeck): Resolved minor conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 34c7b0eeca116ae5dc46804b0f46b8ad47a7d7d7)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>