CHROMIUM: sysrq: add ability for sysrq-x to signal chrome/X
Because we weren't getting reports of actual kernel issues, and
we've already trained some users to hit sysrq-x when the system
appears to freeze, we decided to have sysrq-x try a few other things
in addition to crashing the kernel.  For the first keypress it will
try to find the Chrome browser process, by looking for the one which
has the  session_manager as a parent, and send a SIGABRT to that.  If
there is a second sysrq-x within 5 seconds, it will send a SIGABRT to
the X process, and if there's a third then it will crash the kernel as
before.

BUG=chromium-os:36901
TEST=manual (debugging feature):
 hold down alt-volup then x - chrome should crash
 repeat immediately afterward - session should restart
 repeat and machine should reboot

Change-Id: I127256b7777aee4a4d84a958f2c061d7b6de318b
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39527
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>

[rebase44(filbranden): Squashed and reordered the commits.  Also changed
 initialization of first_jiffies to use INITIAL_JIFFIES and introduced
 time_after() for the time comparison.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
(cherry picked from commit 102e0b92031cd5394237dfb09f8888eda0c17992)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>