UPSTREAM: cgroup: rename cgroupfs_root->number_of_cgroups to ->nr_cgrps and make it atomic_t
root->number_of_cgroups is currently an integer protected with
cgroup_mutex.  Except for sanity checks and proc reporting, the only
place it's used is to check whether the root has any child during
remount; however, this is a bit flawed as the counter is not
decremented when the cgroup is unlinked but when it's released,
meaning that there could be an extended period where all cgroups are
removed but remount is still not allowed because some internal objects
are lingering.  While not perfect either, it'd be better to use
emptiness test on root->top_cgroup.children.

This patch updates cgroup_remount() to test top_cgroup's children
instead, which makes number_of_cgroups only actual usage statistics
printing in proc implemented in proc_cgroupstats_show().  Let's
shorten its name and make it an atomic_t so that we don't have to
worry about its synchronization.  It's purely auxiliary at this point.

Signed-off-by: Tejun Heo <tj@kernel.org>
Acked-by: Li Zefan <lizefan@huawei.com>

BUG=b:29259708
TEST=Built and booted Minnie

(cherry picked from commit 3c9c825b8b50de7dbb015e6bfc04bb2da79364d9)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I583f11a50e3eae3d98da38dfde75617d934d4a89
Reviewed-on: https://chromium-review.googlesource.com/359695
Reviewed-by: Dylan Reid <dgreid@chromium.org>