BACKPORT: cgroup: enable task_cg_lists on the first cgroup mount
Tasks are not linked on their css_sets until cgroup task iteration is
actually used.  This is to avoid incurring overhead on the fork and
exit paths for systems which have cgroup compiled in but don't use it.

This lazy binding also affects the task migration path.  It has to be
careful so that it doesn't link tasks to css_sets when task_cg_lists
linking is not enabled yet.  Unfortunately, this conditional linking
in the migration path interferes with planned migration updates.

This patch moves the lazy binding a bit earlier, to the first cgroup
mount.  It's a clear indication that cgroup is being used on the
system and task_cg_lists linking is highly likely to be enabled soon
anyway through "tasks" and "cgroup.procs" files.

This allows cgroup_task_migrate() to always link @tsk->cg_list.  Note
that it may still race with cgroup_post_fork() but who wins that race
is inconsequential.

While at it, make use_task_css_set_links a bool, add sanity checks in
cgroup_enable_task_cg_lists() and css_task_iter_start(), and update
the former so that it's guaranteed and assumes to run only once.

Signed-off-by: Tejun Heo <tj@kernel.org>
Acked-by: Li Zefan <lizefan@huawei.com>

Conflicts:
	kernel/cgroup.c

- small conflict in cgroup_enable_task_cg_lists() due to patch already
landed in 3.14 proper before cgroup merge.

BUG=b:29259708
TEST=Built and booted Minnie

(cherry picked from commit 56fde9e01de45bcfabbb444d33e8bdd8388d2da0)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I20e7ed48b788d0fc374f04172e9cb6be663d0dcc
Reviewed-on: https://chromium-review.googlesource.com/359700
Reviewed-by: Dylan Reid <dgreid@chromium.org>