UPSTREAM: cgroup: fix the retry path of cgroup_mount()
If we hit the retry path, we'll call parse_cgroupfs_options() again,
but the string we pass to it has been modified by the previous call
to this function.

This bug can be observed by:

  # mount -t cgroup -o name=foo,cpuset xxx /mnt && umount /mnt && \
    mount -t cgroup -o name=foo,cpuset xxx /mnt
  mount: wrong fs type, bad option, bad superblock on xxx,
         missing codepage or helper program, or other error
  ...

The second mount passed "name=foo,cpuset" to the parser, and then it
hit the retry path and call the parser again, but this time the string
passed to the parser is "name=foo".

To fix this, we avoid calling parse_cgroupfs_options() again in this
case.

Signed-off-by: Li Zefan <lizefan@huawei.com>
Signed-off-by: Tejun Heo <tj@kernel.org>

BUG=b:29259708
TEST=Built and booted Minnie

(cherry picked from commit e37a06f10994c2ba86f54d8f96734f2483a869b8)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I1ef11b8849a92d75f05bc383024ba610c411290f
Reviewed-on: https://chromium-review.googlesource.com/359750
Reviewed-by: Dylan Reid <dgreid@chromium.org>