UPSTREAM: cgroup: cgroup_apply_cftypes() shouldn't skip the default hierarhcy
cgroup_apply_cftypes() skip creating or removing files if the
subsystem is attached to the default hierarchy, which led to missing
files in the root of the default hierarchy.

Skipping made sense when the default hierarchy was dummy; however, now
that the default hierarchy is full functional and planned to be used
as the unified hierarchy, it shouldn't be skipped over.

Reported-by: Li Zefan <lizefan@huawei.com>
Signed-off-by: Tejun Heo <tj@kernel.org>
Acked-by: Li Zefan <lizefan@huawei.com>

BUG=b:29259708
TEST=Built and booted Minnie

(cherry picked from commit ea8fd3b47ff4ed4b1b5942bf3e0cb8d8f590ec59)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: Ie88aa6ba14d8ff012ea6444033fa296c5b89a8b8
Reviewed-on: https://chromium-review.googlesource.com/359751
Reviewed-by: Dylan Reid <dgreid@chromium.org>