UPSTREAM: cgroup: implement cgroup->e_csets[]
On the default unified hierarchy, a cgroup may be associated with
csses of its ancestors, which means that a css of a given cgroup may
be associated with css_sets of descendant cgroups.  This means that we
can't walk all tasks associated with a css by iterating the css_sets
associated with the cgroup as there are css_sets which are pointing to
the css but linked on the descendants.

This patch adds per-subsystem list heads cgroup->e_csets[].  Any
css_set which is pointing to a css is linked to
css->cgroup->e_csets[$SUBSYS_ID] through
css_set->e_cset_node[$SUBSYS_ID].  The lists are protected by
css_set_rwsem and will allow us to walk all css_sets associated with a
given css so that we can find out all associated tasks.

Signed-off-by: Tejun Heo <tj@kernel.org>
Acked-by: Li Zefan <lizefan@huawei.com>

BUG=b:29259708
TEST=Built and booted Minnie

(cherry picked from commit 2d8f243a5e6efa57fb7c46fe83fafa45b33d0ec2)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I7c98396cbda74a81065e8b822c56114f39a85168
Reviewed-on: https://chromium-review.googlesource.com/359754
Reviewed-by: Dylan Reid <dgreid@chromium.org>