UPSTREAM: percpu-refcount: rename percpu_ref_tryget() to percpu_ref_tryget_live()
percpu_ref_tryget() is different from the usual tryget semantics in
that it fails if the refcnt is in its dying stage even if the refcnt
hasn't reached zero yet.  We're about to introduce the more
conventional tryget and the current one has only one user.  Let's
rename it to percpu_ref_tryget_live() so that it explicitly signifies
the peculiarities of its semantics.

This is pure rename.

Signed-off-by: Tejun Heo <tj@kernel.org>
Acked-by: Kent Overstreet <kmo@daterainc.com>

BUG=b:29259708
TEST=Built and booted Minnie

(cherry picked from commit 2070d50e1cbe3d7f157cbf8e63279c893f375d7f)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: Ia5d04d631602e3b765dad7fe9281cf5dfcba9cc6
Reviewed-on: https://chromium-review.googlesource.com/359787
Reviewed-by: Dylan Reid <dgreid@chromium.org>