UPSTREAM: cgroup: rename cgroup->dummy_css to ->self and move it to the top
cgroup->dummy_css is used as the placeholder css when performing css
oriended operations on the cgroup.  We're gonna shift more cgroup
management to this css.  Let's rename it to ->self and move it to the
top.

This is pure rename and field relocation.

Signed-off-by: Tejun Heo <tj@kernel.org>
Acked-by: Li Zefan <lizefan@huawei.com>

BUG=b:29259708
TEST=Built and booted Minnie

(cherry picked from commit 9d800df12d31734a6853915e9d2deb5d6747985f)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I4443a7da18a8adafe4a6c3c53dc8fedcf0bd3ce2
Reviewed-on: https://chromium-review.googlesource.com/359811
Reviewed-by: Dylan Reid <dgreid@chromium.org>