FROMLIST: net: Allow xt_owner in any user namespace
Making this work is a little tricky as it really isn't kosher to
change the xt_owner_match_info in a check function.

Without changing xt_owner_match_info we need to know the user
namespace the uids and gids are specified in.  In the common case
net->user_ns == current_user_ns().  Verify net->user_ns ==
current_user_ns() in owner_check so we can later assume it in
owner_mt.

In owner_check also verify that all of the uids and gids specified are
in net->user_ns and that the expected min/max relationship exists
between the uids and gids in xt_owner_match_info.

In owner_mt get the network namespace from the outgoing socket, as this
must be the same network namespace as the netfilter rules, and use that
network namespace to find the user namespace the uids and gids in
xt_match_owner_info are encoded in.  Then convert from their encoded
from into the kernel internal format for uids and gids and perform the
owner match.

BUG=b:29161036
TEST=`/system/bin/iptables -w -A fw_dozable -m owner --uid-owner 0 -j RETURN`
     run from android-sh root shell

Link: https://lists.linuxfoundation.org/pipermail/containers/2014-May/034514.html
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>

Change-Id: Ib9d8b5b3254028124bbdc409f7ba1d9f02c20001
Reviewed-on: https://chromium-review.googlesource.com/350510
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Ricky Zhou <rickyz@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>