CHROMIUM: ext4: Copy symlink data from page
kmapping the page from d_inode->i_mapping like page_getlink() does not
work for ext4 crypto: the data will be put in a scatter gather list and
sent to the encryption engine.
Instead, copy the data to a bounce buffer, to be sure to have a logical
address to put in the sg list.

BUG=chromium:698561
TEST=xfstest generic/068 passes. It used to fail on first symlink.

Change-Id: Ie6283bcaa725f01c5788e7e0b72c73b64fdcff43
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/453125
Reviewed-by: Guenter Roeck <groeck@chromium.org>