UPSTREAM: cgroup: remove CGRP_RELEASABLE flag
We call put_css_set() after setting CGRP_RELEASABLE flag in
cgroup_task_migrate(), but in other places we call it without setting
the flag. I don't see the necessity of this flag.

Moreover once the flag is set, it will never be cleared, unless writing
to the notify_on_release control file, so it can be quite confusing
if we look at the output of debug.releasable.

  # mount -t cgroup -o debug xxx /cgroup
  # mkdir /cgroup/child
  # cat /cgroup/child/debug.releasable
  0   <-- shows 0 though the cgroup is empty
  # echo $$ > /cgroup/child/tasks
  # cat /cgroup/child/debug.releasable
  0
  # echo $$ > /cgroup/tasks && echo $$ > /cgroup/child/tasks
  # cat /proc/child/debug.releasable
  1   <-- shows 1 though the cgroup is not empty

This patch removes the flag, and now debug.releasable shows if the
cgroup is empty or not.

Signed-off-by: Zefan Li <lizefan@huawei.com>
Signed-off-by: Tejun Heo <tj@kernel.org>

BUG=b:29259708
TEST=Built and booted Minnie

(cherry picked from commit a25eb52e81a40e986179a790fbb5a1f02f482b7a)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I18a6ceea176da44ba2afa4aa09678b09e7329b9e
Reviewed-on: https://chromium-review.googlesource.com/359888
Reviewed-by: Dylan Reid <dgreid@chromium.org>