UPSTREAM: kdb: fix incorrect counts in KDB summary command output
The output of KDB 'summary' command should report MemTotal, MemFree
and Buffers output in kB. Current codes report in unit of pages.

A define of K(x) as
is defined in the code, but not used.

This patch would apply the define to convert the values to kB.
Please include me on Cc on replies. I do not subscribe to linux-kernel.

BUG=chrome-os-partner:50495
TEST=Use kgdb / kdb

Change-Id: Ib44455bfe090276bfd24220a5519defcbd80edcc
Signed-off-by: Jay Lan <jlan@sgi.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 146755923262037fc4c54abc28c04b1103f3cc51)
Reviewed-on: https://chromium-review.googlesource.com/329243
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>